requirements.txt
Flask
python-dotenv
pycopg2-binary